+++ lib/fire_flag.dart
@@ -1,4 +1,0 @@
-library fire_flag;
-
-export 'src/feature.dart';
-export 'src/fire_flag.dart';