+++ lib/flag_feature.dart
@@ -1,0 +1,4 @@
+library flag_feature;
+
+export 'src/feature.dart';
+export 'src/feature_flag.dart';